Cleanup plugin API:
- Remove QuiltPluginTask

This was never implemented properly, and I'm unsure if hiding CompletableFuture is even a good idea.
Plus, I haven't actually implemented any multithreading yet - and it will be better to add proper a multithreading API along with the implementation, not a theoretical one beforehand.

- Properly implement TentativeLoadOption.

This part is still untested, but it should work in theory? Either way it no longer requires plugins to deal with tasks.

- Remove GUI reference.

I'm unsure how we should implement plugin gui requests - they could just use the existing gui API instead of needing something specific to plugins.